nua: using isize_t, size_t.
darcs-hash:20060906224631-65a35-9bff3d99f562ad2a6caf88b8ef3c80b4133e1b94.gz

+++ libsofia-sip-ua/nua/nua.c
@@ -1146,7 +1146,7 @@ void nua_signal(nua_t *nua, nua_handle_t *nh, msg_t *msg, int always,
 		tag_type_t tag, tag_value_t value, ...)
 {
   su_msg_r sumsg = SU_MSG_R_INIT;
-  int len, xtra, e_len, l_len = 0, l_xtra = 0;
+  size_t len, xtra, e_len, l_len = 0, l_xtra = 0;
   ta_list ta;
 
   if (nua == NULL || (nua->nua_shutdown_started && event != nua_r_shutdown))
